Implement 12 degrees of movement
Odd numbered movement directions correspond to the existing axial directions (the sides of the hexagons).

Even numbered directions correspond to the vertices of the hexagons: between two sides.

Movement in these even-numbered directions flip-flops between movement to either side of the vertex-direction. For 12 O'Clock, that's alternating between 1 O'Clock and 11 O'Clock.

Turn modes are renamed in terms of "full" and "half" turns. A full turn is clockwise two degrees: for example from 1 O'Clock to 3 O'Clock or from 12 O'Clock to 2 O'Clock. A 180 degree turn is thus termed a "triple turn".

When moving in an even-numbered direction, the Shift R1 and R2 triggers both do the same thing, as there is only one clear option for shifting right relative to such a direction. The same is true for the L1 and L2 shift triggers, respectively